+++ RELEASES.md
@@ -1,6 +1,8 @@
 Version 1.0.0 (2018-??-??)
 ==========================
 
+* `Integer::invert_mut` and `Integer::pow_mod_mut` now return
+  `Result<(), ()>` instead of `bool`.
 * Remove all deprecated items.
 * Remove unsound blanket implementations constrained on
   `T where SmallInteger: Assign<T>` inside `SmallRational` and on